+++ openfisca_us/variables/gov/ssa/ssi/eligibility/is_ssi_aged.py
@@ -1,0 +1,12 @@
+from openfisca_us.model_api import *
+
+
+class is_ssi_aged(Variable):
+    value_type = bool
+    entity = Person
+    label = "Is aged for SSI"
+    definition_period = YEAR
+
+    def formula(person, period, parameters):
+        aged_threshold = parameters(period).ssa.ssi.eligibility.aged_threshold
+        return person("age", period) >= aged_threshold