+++ openfisca_us/variables/gov/ssa/ssi/income/ssi_countable_income.py
@@ -8,11 +8,11 @@ class ssi_countable_income(Variable):
     documentation = "Supplemental Security Income countable income"
     label = "SSI countable income"
     unit = USD
+    reference = "https://www.law.cornell.edu/uscode/text/42/1382a#b"
 
     def formula(person, period, parameters):
-        earned = person("ssi_earned_income", period) / 12
-        unearned = person("ssi_unearned_income", period) / 12
-        deemed = person("ssi_deemed_income", period) / 12
+        earned = person("ssi_earned_income", period) / MONTHS_IN_YEAR
+        unearned = person("ssi_unearned_income", period) / MONTHS_IN_YEAR
         # Deduct exclusions.
         exclusions = parameters(period).ssa.ssi.income.exclusions
         # Subtract from unearned income first.
@@ -25,6 +25,4 @@ class ssi_countable_income(Variable):
         earned_after_flat_exclusion = max_(earned - earned_exclusion, 0)
         countable_earned_share = 1 - exclusions.earned_share
         countable_earned = earned_after_flat_exclusion * countable_earned_share
-        return (
-            countable_unearned + countable_earned + deemed
-        ) * MONTHS_IN_YEAR
+        return (countable_unearned + countable_earned) * MONTHS_IN_YEAR