+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.66.0] - 2022-05-19 12:47:08
+
+### Added
+
+- MaritalUnit entity.
+- Massachusetts state supplement.
+
+## [0.65.0] - 2022-05-19 11:54:27
+
+### Added
+
+- WIC takeup and nutritional risk imputations.
+
+## [0.64.1] - 2022-05-17 22:49:48
+
+### Fixed
+
+- Corrected EITC phase-out start values for 2020 and 2021.
+
 ## [0.64.0] - 2022-05-16 20:12:47
 
 ### Changed
@@ -844,6 +863,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+[0.66.0]: https://github.com/PolicyEngine/openfisca-us/compare/0.65.0...0.66.0
+[0.65.0]: https://github.com/PolicyEngine/openfisca-us/compare/0.64.1...0.65.0
+[0.64.1]: https://github.com/PolicyEngine/openfisca-us/compare/0.64.0...0.64.1
 [0.64.0]: https://github.com/PolicyEngine/openfisca-us/compare/0.63.0...0.64.0
 [0.63.0]: https://github.com/PolicyEngine/openfisca-us/compare/0.62.3...0.63.0
 [0.62.3]: https://github.com/PolicyEngine/openfisca-us/compare/0.62.2...0.62.3